site: new simpler docs page
This docs page is less verbose, and should make it easier to know what
you should read next, and what you don't need to read.

The reference docs are still in flux, with three alternatives presented:
1. the old manual, 2. the new individual man pages, and 3. the new
manual which combines the man pages. At this point I think that 2 will
get top billing, 3 will also be provided, and 1 will be dropped when 2/3
are a bit better.