wiki: move FAQ back to markdown, add an explanatory note.
The note:

I'd rather maintain this in org mode.
As of 20190418 github (org-ruby ?) doesn't render org well;
it converts to "smart" quotes/hyphens, and misrenders =--drop=.
This would make our docs look bad when viewed on github.
They are intended to be viewed on hledger.org, which avoids these problems.
Still, keeping it in markdown for now.